+++ client/src/components/HeroImage/HeroImage.styles.js
