+++ client/src/components/HeroImage/index.js
