+++ client/src/components/Movie.js
