+++ client/src/components/MovieInfo/MovieInfo.styles.js
