+++ client/src/components/MovieInfo/index.js
