+++ client/src/components/MovieInfoBar/MovieInfoBar.styles.js
