+++ client/src/components/MovieInfoBar/index.js
