+++ client/src/components/Rate/index.js
