+++ client/src/components/SearchBar/SearchBar.styles.js
