+++ client/src/components/Signup/index.js
@@ -3,10 +3,10 @@ import { useNavigate } from "react-router-dom";
 import API from "../../API";
 
 import Button from "../Button";
-import { Wrapper } from "./Login.styles";
+import { Wrapper } from "./Signup.styles";
 import { Context } from "../../context";
 
-const Login = () => {
+const Signup = () => {
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
   const [error, setError] = useState(false);
@@ -44,6 +44,7 @@ const Login = () => {
   return (
     <Wrapper>
       {error && <div className="error">Oops! Something went wrong!</div>}
+      <h2>Sign Up</h2>
       <label htmlFor="">Username:</label>
       <input
         type="text"
@@ -58,9 +59,9 @@ const Login = () => {
         name="password"
         onChange={handleInput}
       />
-      <Button text="Login" callback={handleSubmit} />
+      <Button text="Signup" callback={handleSubmit} />
     </Wrapper>
   );
 };
 
-export default Login;
+export default Signup;
