+++ client/src/components/Thumb/Thumb.styles.js
