+++ client/src/components/Thumb/index.js
